requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
pyyaml==6.0.1
ultralytics==8.1.0
opencv-python==4.8.1.78
deepface==0.0.79
tf-keras==2.18.0
pytesseract==0.3.10
pyttsx3==2.90
SpeechRecognition==3.10.0
vosk==0.3.45
google-generativeai==0.3.2
numpy>=1.26.0,<2.0.0
pillow==10.1.0
eventlet==0.33.3
pyaudio==0.2.14